Tweaks and more tests

+++ keanu-project/src/main/java/io/improbable/keanu/util/graph/GraphNode.java
@@ -5,5 +5,5 @@ import java.util.Map;
 public interface GraphNode {
     Map<String,String> getDetails();
 
-    int getIndex();
+    long getIndex();
 }
